[ADAM-356] Refactor packages for org.bdgenomics.adam.rdd contents.
