+++ adam-core/src/main/scala/org/bdgenomics/adam/rdd/pileup/ADAMPileupRDDFunctions.scala
@@ -17,24 +17,10 @@
  */
 package org.bdgenomics.adam.rdd.pileup
 
-import java.util.logging.Level
-import org.apache.avro.specific.SpecificRecord
 import org.apache.spark.Logging
-import org.apache.spark.SparkContext._
 import org.apache.spark.rdd.RDD
 import org.bdgenomics.adam.models._
-import org.bdgenomics.adam.rdd.ADAMContext._
-import org.bdgenomics.adam.util.{
-  HadoopUtil,
-  ParquetLogger
-}
 import org.bdgenomics.formats.avro._
-import parquet.avro.AvroParquetOutputFormat
-import parquet.hadoop.ParquetOutputFormat
-import parquet.hadoop.metadata.CompressionCodecName
-import parquet.hadoop.util.ContextUtil
-import scala.math.max
-import scala.Some
 
 class ADAMPileupRDDFunctions(rdd: RDD[Pileup]) extends Serializable with Logging {
   /**
